reviews/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Cors==5.0.0
Flask-HTTPAuth==4.8.0
flask-profiler==1.8.1
mysql-connector-python==8.0.33
PyJWT==2.7.0
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.28.0
pymysql

